Rephrase the banner for the state of the project
Previous wording incorrectly used "have been" for a third-person singular subject, instead of "has been". Since the present perfect refers to the recent past, it's preferable to either use the past simple (we published an update), or to just avoid usage of it altogether, which is what this commit does.